requirements.txt
Django==3.1.12
djangorestframework==3.12.4
djongo==1.3.7
python-dotenv==1.0.1
Requests==2.32.3
web3==7.6.0
